backend/requirements.txt
flask
flask-cors
python-dotenv
requests
flask-sqlalchemy
pyjwt
slackclient
pandas
scikit-learn